requirements.txt
langchain==0.0.135
spotipy==2.23.0
python-dotenv==1.0.0 